+++ src/nespresso/core/logs/settings.py
@@ -28,12 +28,6 @@ _FILE_FORMAT = JsonFormatter(
 )
 
 
-class EscapeNewlinesFilter(logging.Filter):
-    def filter(self, record: logging.LogRecord) -> bool:
-        record.msg = record.getMessage().replace("\n", "\\n")
-        return True
-
-
 class AiogramFilter(logging.Filter):
     def __init__(self, level: int = 100):  # 100 - block of any logs
         super().__init__()
@@ -81,8 +75,6 @@ def CreateConsoleHandler(
     handler = StreamHandler()
     handler.setLevel(level)
     handler.setFormatter(_CONSOLE_FORMAT)
-
-    handler.addFilter(EscapeNewlinesFilter())
 
     if filters:
         for filt in filters: