+++ src/nespresso/db/models/message.py
@@ -1,7 +1,7 @@
-from datetime import UTC, datetime
+from datetime import datetime
 from enum import Enum
 
-from sqlalchemy import BigInteger, DateTime, Enum as SqlEnum, String
+from sqlalchemy import BigInteger, DateTime, Enum as SqlEnum, String, text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from nespresso.db.base import Base
@@ -20,8 +20,11 @@ class Message(Base):
     )
 
     time: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), default=datetime.now(tz=UTC)
+        DateTime(timezone=True),
+        server_default=text("CURRENT_TIMESTAMP"),
+        nullable=False,
     )
+
     chat_id: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
     side: Mapped[MessageSide] = mapped_column(SqlEnum(MessageSide), nullable=False)
     text: Mapped[str] = mapped_column(String, nullable=False)