+++ CHANGELOG.md
@@ -21,6 +21,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## [0.1.0] - 2025-05-24 - Vlad
+### Updated
+- Simplify UserService
+
 ## [0.1.0] - 2025-05-22 - Vlad
 ### Added
 - Send notification to user in case of error